サウンドを修正
