+++ team_GL/team_GL.vcxproj
@@ -89,6 +89,7 @@
     <ClCompile Include="scene.cpp" />
     <ClCompile Include="scene2D.cpp" />
     <ClCompile Include="scene3D.cpp" />
+    <ClCompile Include="soundal.cpp" />
     <ClCompile Include="texture.cpp" />
     <ClCompile Include="title.cpp" />
     <ClCompile Include="Vector.cpp" />
@@ -107,6 +108,7 @@
     <ClInclude Include="scene.h" />
     <ClInclude Include="scene2D.h" />
     <ClInclude Include="scene3D.h" />
+    <ClInclude Include="soundal.h" />
     <ClInclude Include="texture.h" />
     <ClInclude Include="title.h" />
     <ClInclude Include="Vector.h" />
